Clean up shared command helpers
